scripts/inference/tvm/requirements.txt
tvm==0.10.0
pyyaml
numpy
opencv-python-headless
